# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MikroTik Centralized Fleet Management API
  version: 1.20.0
...
servers:
- url: /api/v1
paths:
...
  /telemetry/ingest:
    post:
      summary: Persist a batch of raw telemetry events
      description: Persists raw interface counters, RADIUS accounting or IPFIX flow events in the raw inbox. This endpoint does not claim that a RouterOS collector is connected and does not run analytics.
      tags:
      - Telemetry
      parameters:
      - in: header
        name: x-actor-permissions
        required: false
        description: Temporary AuthStubGuard permission header; production authentication is not implemented in this phase.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryIngestRequest'
      responses:
        '201':
          description: Events persisted or acknowledged as idempotent duplicates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryIngestResponse'
        '400':
          description: Invalid telemetry payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: IDEMPOTENCY_KEY_REUSED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: TELEMETRY_IDENTITY_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /tenants:
    get:
      summary: List tenant (phang, dung parent_id de tu dung cay o frontend)
      tags:
      - Tenants
      parameters:
      - *id014
      - in: query
        name: parent_id
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
    post:
      summary: Tao tenant moi
      tags:
      - Tenants
      parameters:
      - *id014
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
  /tenants/{tenantId}:
    get:
      summary: Chi tiet tenant
      tags:
      - Tenants
      parameters:
      - *id014
      - in: path
        name: tenantId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
        '404': *id018
    patch:
      summary: Cap nhat tenant
      tags:
      - Tenants
      parameters:
      - *id014
      - in: path
        name: tenantId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
    delete:
      summary: Xoa tenant (soft delete). Chan neu con tenant con hoac subscriber (TENANT_HAS_DEPENDENTS).
      tags:
      - Tenants
      parameters:
      - *id014
      - in: path
        name: tenantId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
        '403': *id015
        '404': *id018
        '409':
          description: TENANT_HAS_DEPENDENTS -- con tenant con hoac subscriber dang thuoc tenant nay
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
...
  /subscribers:
    get:
      summary: List subscriber PPPoE/Hotspot (loc tenant_id/status, tim theo username qua q)
      tags:
      - Subscribers
      parameters:
      - *id014
      - in: query
        name: tenant_id
        schema:
          type: string
          format: uuid
      - in: query
        name: nas_device_id
        schema:
          type: string
          format: uuid
        description: Loc theo NAS (MikroTik router) dang gan cho subscriber -- dung o Device Detail de xem user gan vao thiet bi.
      - in: query
        name: status
        schema:
          type: string
          enum: &id100
          - ACTIVE
          - SUSPENDED
          - EXPIRED
      - in: query
        name: q
        schema:
          type: string
        description: Tim theo username (chua dung, khong phan biet hoa/thuong)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscriber'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
    post:
      summary: Tao subscriber moi (gan tenant + goi cuoc, tuy chon NAS)
      tags:
      - Subscribers
      parameters:
      - *id014
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Subscriber'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
  /subscribers/bulk-assign:
    post:
      summary: Gan hang loat goi cuoc va/hoac NAS cho nhieu subscriber cung luc
      description: 'De trong package_id hoac nas_device_id = giu nguyen gia tri hien tai cua tung subscriber (README thiet ke muc 7 buoc 2).'
      tags:
      - Subscribers
      parameters:
      - *id014
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberBulkAssignRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      updated_count:
                        type: integer
                      subscribers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Subscriber'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
  /subscribers/{subscriberId}:
    get:
      summary: Chi tiet subscriber
      tags:
      - Subscribers
      parameters:
      - *id014
      - in: path
        name: subscriberId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Subscriber'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
        '404': *id018
    patch:
      summary: Cap nhat subscriber (doi goi/NAS/trang thai/han dung)
      tags:
      - Subscribers
      parameters:
      - *id014
      - in: path
        name: subscriberId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Subscriber'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
    delete:
      summary: Xoa subscriber (soft delete)
      tags:
      - Subscribers
      parameters:
      - *id014
      - in: path
        name: subscriberId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
        '403': *id015
        '404': *id018
  /subscribers/{subscriberId}/password:
    post:
      summary: Cap mat khau RADIUS moi cho subscriber
      description: Sinh mat khau that ngau nhien, bam scrypt luu lai (backend/src/libs/password-hash/), tra plaintext DUY NHAT 1 LAN. RADIUS Access-Request (PAP) ke tiep tu router dung duoc ngay.
      tags:
      - Subscribers
      parameters:
      - *id014
      - in: path
        name: subscriberId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SubscriberPasswordResponse'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
        '404': *id018
    delete:
      summary: Thu hoi mat khau RADIUS cua subscriber
      description: Xoa password_hash/password_issued_at -- Access-Request ke tiep voi mat khau cu se bi tu choi (Access-Reject).
      tags:
      - Subscribers
      parameters:
      - *id014
      - in: path
        name: subscriberId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Revoked
        '403': *id015
        '404': *id018
...
  /telemetry/normalize:
    post:
      summary: Manually catch up / reprocess pending raw telemetry normalization
      description: Stand-in for the NATS consumer worker described in ADR-03 (not implemented in this phase — see docs/backend/02-DATABASE_DESIGN.md §8). Ingest already normalizes synchronously; this retries anything left with processed_at IS NULL.
      tags:
      - Telemetry
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 1000
          default: 200
      - in: query
        name: source
        required: false
        schema:
          $ref: '#/components/schemas/TelemetrySource'
      - in: header
        name: x-actor-permissions
        required: false
        description: Temporary AuthStubGuard permission header; production authentication is not implemented in this phase.
        schema:
          type: string
      responses:
        '200':
          description: Normalization sweep result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryNormalizeResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
    BaseMeta:
      type: object
      properties:
        request_id:
          type: string
        generated_at:
          type: string
          format: date-time
        data_from:
          type: string
          format: date-time
        data_to:
          type: string
          format: date-time
        data_freshness_seconds:
          type: integer
          nullable: true
        freshness_by_source:
          type: object
          additionalProperties:
            type: integer
            nullable: true
        page:
          type: object
          nullable: true
          description: Chi xuat hien tren cac endpoint list co phan trang (cursor hoac offset).
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
            cursor_next:
              type: string
              nullable: true
            has_more:
              type: boolean
        warnings:
          type: array
          nullable: true
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              details:
                type: object
                additionalProperties: true
...
    ErrorEnvelope:
      type: object
      properties:
        data:
          nullable: true
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          $ref: '#/components/schemas/ErrorDetails'
    TelemetrySource:
      type: string
      enum:
      - INTERFACE_COUNTER
      - RADIUS_ACCOUNTING
      - IPFIX_FLOW
    TelemetryIdentity:
      type: object
      properties:
        ship_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        interface_id:
          type: string
          format: uuid
        user_identity:
          type: string
        user_identity_type:
          type: string
          enum:
          - USERNAME
          - IP
          - MAC
          - SESSION_ID
          - UNKNOWN
        source_device_ref:
          type: string
        source_interface_ref:
          type: string
    TelemetryEvent:
      type: object
      required:
      - source
      - idempotency_key
      - observed_at
      properties:
        source:
          $ref: '#/components/schemas/TelemetrySource'
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 255
        source_event_id:
          type: string
        observed_at:
          type: string
          format: date-time
        identity:
          $ref: '#/components/schemas/TelemetryIdentity'
        payload:
          type: object
          additionalProperties: true
          description: Source-specific raw JSON payload. At least payload or raw_reference is required.
        raw_reference:
          type: string
          description: Reference to immutable object storage when payload is stored outside the API.
        metadata:
          type: object
          additionalProperties: true
      description: 'Source-specific validation is enforced by the backend: interface counters require interface_name/rx_bytes/tx_bytes; RADIUS requires acct_status_type/acct_session_id; IPFIX requires bytes.'
    TelemetryIngestRequest:
      type: object
      required:
      - events
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/TelemetryEvent'
    TelemetryIngestResult:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        source:
          $ref: '#/components/schemas/TelemetrySource'
        idempotency_key:
          type: string
        status:
          type: string
          enum:
          - ACCEPTED
          - DUPLICATE
        observed_at:
          type: string
          format: date-time
        normalized:
          type: boolean
          nullable: true
          description: Outcome of NormalizationService.processEvent() for this event on THIS ingest call. null for DUPLICATE events (already normalized on first ingest, or intentionally not re-run).
    TelemetryIngestResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accepted_count:
              type: integer
            duplicate_count:
              type: integer
            events:
              type: array
              items:
                $ref: '#/components/schemas/TelemetryIngestResult'
            persistence:
              type: string
              enum:
              - raw_telemetry_events
            analytics_processed:
              type: boolean
              description: true only when EVERY accepted event in this batch was actually normalized into interface_counter_*/radius_*/ipfix_flow_records (see NormalizationService). Never fabricated as true; a partial or failed normalization leaves this false and the event stays available for POST /telemetry/normalize to retry.
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...
    ErrorDetails:
      type: object
      description: Body loi chuan hoa -- AGENT_COLLABORATION.md Sec.5. Khong bao gio chua stack trace hay loi driver that.
      properties:
        code:
          type: string
          enum:
          - VALIDATION_FAILED
          - INVALID_TIME_RANGE
          - INSUFFICIENT_DATA
          - RECONCILIATION_UNAVAILABLE
          - UNAUTHENTICATED
          - FORBIDDEN
          - SCOPE_FORBIDDEN
          - AREA_NOT_FOUND
          - SHIP_NOT_FOUND
          - DEVICE_NOT_FOUND
          - INTERFACE_NOT_FOUND
          - ZONE_NOT_FOUND
          - ENDPOINT_NOT_FOUND
          - ALERT_NOT_FOUND
          - SERVICE_NOT_FOUND
          - RESOURCE_CONFLICT
          - AREA_HAS_SHIPS
          - SHIP_HAS_DEVICES
          - INTERFACE_DOUBLE_COUNT
          - LAST_HEALTHY_ENDPOINT
          - ENDPOINT_ALREADY_EXISTS
          - DATABASE_UNAVAILABLE
          - RATE_LIMITED
          - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
        retryable:
          type: boolean
        retry_after_seconds:
          type: integer
          nullable: true
      required:
      - code
      - message
      - retryable
...
    Tenant:
      type: object
      description: Tenant/dai ly phan cap (nha cung cap -> cong ty/dai ly -> chi nhanh). parent_id null = tenant goc.
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        contact_phone:
          type: string
          nullable: true
        contact_email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        tax_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TenantCreateRequest:
      type: object
      required:
      - code
      - name
      properties:
        parent_id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        contact_phone:
          type: string
          nullable: true
        contact_email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        tax_id:
          type: string
          nullable: true
    TenantUpdateRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        contact_phone:
          type: string
          nullable: true
        contact_email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        tax_id:
          type: string
          nullable: true
...
    Subscriber:
      type: object
      description: Subscriber PPPoE/Hotspot. NAS = MikroTik router (devices.id). Mat khau RADIUS that (hash scrypt) khong bao gio tra ve -- chi bao da cap hay chua qua password_configured.
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        username:
          type: string
        auth_type:
          type: string
          enum: &id102
          - PPPOE
          - HOTSPOT
        nas_device_id:
          type: string
          format: uuid
          nullable: true
        package_id:
          type: string
          format: uuid
        status:
          type: string
          enum: *id100
        quota_used_bytes:
          type: integer
          format: int64
        expires_at:
          type: string
          format: date-time
        password_configured:
          type: boolean
          description: true neu subscriber da duoc cap mat khau RADIUS that qua POST .../password.
        password_issued_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubscriberCreateRequest:
      type: object
      required:
      - tenant_id
      - username
      - auth_type
      - package_id
      - expires_at
      properties:
        tenant_id:
          type: string
          format: uuid
        username:
          type: string
        auth_type:
          type: string
          enum: *id102
        nas_device_id:
          type: string
          format: uuid
          nullable: true
        package_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
    SubscriberUpdateRequest:
      type: object
      properties:
        auth_type:
          type: string
          enum: *id102
        nas_device_id:
          type: string
          format: uuid
          nullable: true
        package_id:
          type: string
          format: uuid
        status:
          type: string
          enum: *id100
        expires_at:
          type: string
          format: date-time
    SubscriberBulkAssignRequest:
      type: object
      required:
      - subscriber_ids
      description: 'De trong package_id/nas_device_id = giu nguyen gia tri hien tai cua tung subscriber.'
      properties:
        subscriber_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        package_id:
          type: string
          format: uuid
        nas_device_id:
          type: string
          format: uuid
    SubscriberPasswordResponse:
      type: object
      description: Mat khau RADIUS that cua subscriber -- CHI xuat hien trong response nay, khong bao gio doc lai duoc sau do.
      properties:
        password:
          type: string
          description: Plaintext, hien mot lan duy nhat.
        issued_at:
          type: string
          format: date-time
...
    TelemetryNormalizeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            scanned:
              type: integer
            normalized:
              type: integer
            failed:
              type: integer
            details:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  ok:
                    type: boolean
                  note:
                    type: string
                    nullable: true
        meta:
          type: object
          additionalProperties: true
        error:
          nullable: true
      description: Manual catch-up/reprocess trigger — stand-in for the NATS consumer worker (ADR-03). Ingest already normalizes synchronously; this is for events left pending after a transient failure.
...